PT-8276: External sign in logic is moved to the service (#2552)
* External sign in logic is moved to the service

* External sign in logic is moved to the service

* Workflow update
