+++ src/VirtoCommerce.Platform.Web/Startup.cs
@@ -435,6 +435,8 @@ namespace VirtoCommerce.Platform.Web
             services.AddOptions<LoginPageUIOptions>().Bind(loginPageUIOptions);
             services.AddDatabaseDeveloperPageExceptionFilter();
             services.AddHttpClient();
+
+            services.AddTransient<IExternalSigninService, ExternalSigninService>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
